skip docs gen and check on empty example folders

+++ avm_scripts/docs-check.sh
@@ -2,6 +2,10 @@
 
 check_docs () {
   local dir=$1
+  if ! ls "$dir"/*.tf "$dir"/*.tf.json 1> /dev/null 2>&1; then
+    echo "===> Skipping $dir as it contains no .tf or .tf.json files"
+    return
+  fi
   echo "===> Generating documentation in $dir"
 	cp "$dir/README.md" "$dir/README-generated.md"
   rm -f "$dir/.terraform.lock.hcl"
